+++ packages/core/CHANGELOG.md
@@ -9,6 +9,7 @@ Versioning].
 ## [Unreleased]
 
 - Fix ignored options in `BuiltInManglers` and `RecommendedManglers`.
+- Fix unintentional mangling in HTML attributes.
 
 ## [0.1.22] - 2021-06-15