+++ packages/core/CHANGELOG.md
@@ -8,7 +8,8 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- Fix ignored options in `BuiltInManglers` and `RecommendedManglers`.
+- Fix unintentional mangling in HTML attributes.
 
 ## [0.1.22] - 2021-06-15